LibHTML: Implement "text-align: justify"
In order for this to work nicely, I made the line box classes use float
instead of int for its geometry information.

Justification works by distributing all of the whitespace on the line
(including the trailing whitespace before the line break) evenly across
the spaces in-between words.

We should probably use floating point (or maybe fixed point?) for all
the layout metrics stuff. But one thing at a time. :^)

+++ Libraries/LibHTML/Layout/LineBox.h
@@ -7,7 +7,7 @@ class LineBox {
 public:
     LineBox() {}
 
-    int width() const { return m_width; }
+    float width() const { return m_width; }
 
     void add_fragment(const LayoutNode& layout_node, int start, int length, int width, int height);
 
@@ -16,5 +16,5 @@ public:
 
 private:
     Vector<LineBoxFragment> m_fragments;
-    int m_width { 0 };
+    float m_width { 0 };
 };
